PythonProject/requirements.txt
fastapi
uvicorn
transformers
torch
pdfplumber
pytesseract
pillow
python-docx
langdetect
nltk
sentencepiece
